Modify: moving dto conversion logic from the controller layer to the service layer.
